+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 ### v0.1.0
 Initial release.
+
+### v0.1.1
+- Added ability to grouping copies from selected items as bundle group.
+- Fixed issue in match transforms with some herarchies by forcing to use topnode only